requirements.txt
asgiref
Django
django-cors-headers
django-environ
django-filter
djangorestframework
Markdown
pillow
psycopg2-binary
sqlparse
tzdata